fix: INCT-75, form title size, refactor scss
