+++ src/shared/ui/InputField/InputField.tsx
@@ -4,13 +4,13 @@ import { default as cn } from 'classnames'
 
 import { IField } from '../../types'
 
-import styles from './InputField.module.scss'
+import styles from './InputField.module.css'
 
 export const InputField = forwardRef<HTMLInputElement, IField>(
   ({ placeholder, helperText, label, type = 'text', style, ...rest }, ref) => {
     return (
       <div className={styles.wrapper} style={style}>
-        <label>{label}</label>
+        <label className={styles.label}>{label}</label>
         <div className={helperText ? '' : 'mb-4'}>
           <input
             autoComplete="off"
