+++ src/shared/ui/PasswordField/PasswordField.module.css
@@ -1,0 +1,11 @@
+.label {
+    @apply block mb-1 text-light-900 text-sm
+}
+
+.input {
+    @apply w-full bg-dark-500 border-2 p-2 pr-10 rounded text-light-100 outline-none
+}
+
+.input::placeholder {
+    @apply text-light-900
+}
