--- src/shared/ui/PasswordField/PasswordField.module.scss
@@ -1,25 +1,0 @@
-.wrapper {
-
-    >label {
-        @apply block mb-1 text-light-900 text-sm
-    }
-
-    >div {
-        @apply relative;
-
-
-
-        >span {
-            @apply absolute top-1/2 right-3 -translate-y-1/2 fill-light-100 cursor-pointer
-        }
-    }
-
-    .input {
-        @apply w-full bg-dark-500 border-2 p-2 pr-10 rounded text-light-100 outline-none
-    }
-
-    ::placeholder {
-        @apply text-light-900
-    }
-
-}
