+++ src/shared/ui/PasswordField/PasswordField.tsx
@@ -4,7 +4,7 @@ import { default as cn } from 'classnames'
 
 import { IField } from '../../types'
 
-import styles from './PasswordField.module.scss'
+import styles from './PasswordField.module.css'
 
 import { EyeOffOutlineIcon, EyeOutlineIcon } from '@/shared/assets'
 
@@ -18,9 +18,9 @@ export const PasswordField = forwardRef<HTMLInputElement, IField>(
     }
 
     return (
-      <div className={styles.wrapper} style={style}>
-        <label>{label}</label>
-        <div className={helperText ? '' : 'mb-4'}>
+      <div style={style}>
+        <label className={styles.label}>{label}</label>
+        <div className={`${helperText ? '' : 'mb-4'} relative`}>
           <input
             autoComplete="off"
             ref={ref}
@@ -29,7 +29,10 @@ export const PasswordField = forwardRef<HTMLInputElement, IField>(
             placeholder={placeholder}
             className={cn(styles.input, helperText ? 'border-danger-500' : 'border-dark-100 ')}
           />
-          <span onClick={togglePassword}>
+          <span
+            className="absolute top-1/2 right-3 -translate-y-1/2 fill-light-100 cursor-pointer"
+            onClick={togglePassword}
+          >
             {type === 'password' ? <EyeOutlineIcon /> : <EyeOffOutlineIcon />}
           </span>
         </div>
